confirmation needed for remove previous output

+++ Preprocessing/ROI2Col.sh
@@ -6,6 +6,13 @@ ROIPath=${ROIPath:-'../../personnel-records/1956/seg/ROI'}
 
 OutputPath=${OutputPath:-'../../personnel-records/1956/seg/col_rect'}
 
-rm $OutputPath --recursive
+read -p "Do you want to remove previous output in $OutputPath? (y/n) " -n 1 -r
+echo -e "\n"
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+    echo "removing $OutputPath"
+
+    rm $OutputPath --recursive
+fi
 
 python3 ROI2Col.py --imgdir=$ImgPath --ROIdir=$ROIPath --outputdir=$OutputPath 2>&1 | tee log_ROI.txt
